Refactoring: moved http stuff to GPCloudHttpImpl
